+++ common/src/main/java/ch/algotrader/service/FutureService.java
@@ -32,7 +32,7 @@ public interface FutureService {
     /**
      * Creates the missing Part of the Futures Chain defined by the {@code futureFamilyId}.
      */
-    public void createDummyFutures(int futureFamilyId);
+    public void createDummyFutures(long futureFamilyId);
 
     /**
      * Get the first Future with {@code expirationDate} after the specified {@code
@@ -41,7 +41,7 @@ public interface FutureService {
      * simulateFuturesByGenericFutures} is configured, a dummy Future will be created if none was
      * found.
      */
-    public Future getFutureByMinExpiration(int futureFamilyId, Date targetExpirationDate);
+    public Future getFutureByMinExpiration(long futureFamilyId, Date targetExpirationDate);
 
     /**
      * Gets a Future by its {@link ch.algotrader.entity.security.FutureFamily} and {@code expirationDate}.
@@ -49,7 +49,7 @@ public interface FutureService {
      * simulateFuturesByGenericFutures} is configured, a dummy Future will be created if none was
      * found.
      */
-    public Future getFutureByExpiration(int futureFamilyId, Date expirationDate);
+    public Future getFutureByExpiration(long futureFamilyId, Date expirationDate);
 
     /**
      * Gets a Future by its {@link ch.algotrader.entity.security.FutureFamily}, that is {@code n} durations after the specified
@@ -58,13 +58,13 @@ public interface FutureService {
      * simulateFuturesByGenericFutures} is configured, a dummy Future will be created if none was
      * found.
      */
-    public Future getFutureByDuration(int futureFamilyId, Date targetDate, int duration);
+    public Future getFutureByDuration(long futureFamilyId, Date targetDate, int duration);
 
     /**
      * Gets all Futures with {@code expirationDate} after the specified {@code targetExpirationDate}
      * and {@code futureFamilyId}. The returned Futures are sorted with their {@code expirationDate}
      * in ascending order.
      */
-    public List<Future> getFuturesByMinExpiration(int futureFamilyId, Date minExpirationDate);
+    public List<Future> getFuturesByMinExpiration(long futureFamilyId, Date minExpirationDate);
 
 }
