+++ common/src/main/java/ch/algotrader/vo/MarketDataEventVO.java
@@ -35,7 +35,7 @@ public class MarketDataEventVO implements Serializable {
     /**
      * The Id of the Security associated with this MarketDataEventVO
      */
-    private int securityId;
+    private long securityId;
 
     /**
      * boolean setter for primitive attribute, so we can tell if it's initialized
@@ -155,7 +155,7 @@ public class MarketDataEventVO implements Serializable {
      * The Id of the Security associated with this MarketDataEventVO
      * @return securityId int
      */
-    public int getSecurityId() {
+    public long getSecurityId() {
 
         return this.securityId;
     }
@@ -164,7 +164,7 @@ public class MarketDataEventVO implements Serializable {
      * The Id of the Security associated with this MarketDataEventVO
      * @param value int
      */
-    public void setSecurityId(final int value) {
+    public void setSecurityId(final long value) {
 
         this.securityId = value;
         this.setSecurityId = true;
